# admin-api/src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Unchain Admin API
  description: |
    Comprehensive OpenAPI specification for the Unchain Admin API.

    This API controls the configuration of the Unchain Feature Management platform, 
    including Features, Strategies, Projects, Environments, Context Fields, 
    Change Requests, Segments, FeatureTags, and Dependencies.

    **Base URL**: `<your-unchain-instance>/api/admin`
  version: 4.0.0
  contact:
    name: Philipp Walther
servers:
  - url: http://localhost:8080
    description: Local Instance
security:
  - OidcAuth: []
...
paths:
...
  /tags:
    get:
      operationId: listTags
      summary: List all tags
      tags: [FeatureTag]
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeatureTag"
    post:
      operationId: createTag
      summary: Create a new tag
      tags: [FeatureTag]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureTag"
      responses:
        "201":
          description: Tag created
...
  /features/{featureName}/tags:
    post:
      operationId: addTagToFeature
      summary: Add a tag to a feature
      tags: [FeatureTag]
      parameters:
        - $ref: "#/components/parameters/featureName"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureTag"
      responses:
        "201":
          description: Tag added to feature

# =========================================================================
# COMPONENTS
# =========================================================================
components:
  securitySchemes:
    OidcAuth:
      type: openIdConnect
      openIdConnectUrl: https://server.redmoon.ch/auth/oidc/.well-known/openid-configuration
  parameters:
...
    featureName:
      name: featureName
      in: path
      required: true
      schema:
        type: string
      description: The name of the feature toggle
...
  schemas:
...
    FeatureTag:
      type: object
      required:
        - type
        - value
      properties:
        type:
          type: string
          example: "simple"
        value:
          type: string
          example: "payment-team"
...